fix(db): resolve Clang-Tidy warnings and errors in tests and core logic

+++ tests/unit/test_api_safety.cpp
@@ -65,7 +65,7 @@ TEST_F(ApiServerSafetyTest, InferenceValidatesCount) {
     body["model_run_id"] = "test_model";
     std::vector<nlohmann::json> samples;
     samples.reserve(1001);
-    for(int i=0; i<1001; ++i) {
+    for(size_t i=0; i<1001; ++i) {
         samples.emplace_back(nlohmann::json{{"cpu_usage", 0.5}});
     }
     body["samples"] = samples;
